requirements.txt
IMDbPY==2022.7.9
IMDbPY==2022.7.9
joblib==1.5.1
pandas==2.3.2
passlib==1.7.4
prophet==1.1.7
protobuf==6.32.0
python-dotenv==1.1.1
pytrends==4.9.2
scikit_learn==1.7.1
sentence_transformers==5.1.0
SQLAlchemy==2.0.43
streamlit==1.48.1
torch==2.8.0
google-genai
bcrypt
Flask==2.3.3
pandas==2.0.3
sqlalchemy==2.0.20
passlib==1.7.4
werkzeug==2.3.7
gunicorn==21.2.0